added

+++ jonas-schemmedman/ultimate-react-course-main/fast-pizza-co/src/features/cart/cartSlice.js
@@ -25,9 +25,7 @@ const cartSlice = createSlice({
       const item = state.cart.find((item) => item.pizzaId === action.payload);
       item.quantity--;
       item.totalPrice = item.quantity * item.unitPrice;
-      if (item.quantity === 0) {
-        cartSlice.caseReducers.deleteItem(state, action);
-      }
+      if (item.quantity === 0) cartSlice.caseReducers.deleteItem(state, action);
     },
     clearCart(state) {
       state.cart = [];
